Retouche makefile et qlqs fichiers

+++ srcs/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: mmanley <mmanley@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/08 14:33:57 by mmanley           #+#    #+#             */
-/*   Updated: 2017/11/24 19:36:36 by mmanley          ###   ########.fr       */
+/*   Updated: 2018/05/12 13:44:19 by mmanley          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,10 +26,10 @@ static int				white_space_counter(char *tmp, int i, int *sign)
 	return (i);
 }
 
-int						ft_atoi(const char *str)
+long long						ft_atoi(const char *str)
 {
 	int					i;
-	unsigned long long	a;
+	long long			a;
 	int					sign;
 	char				*tmp;
 
@@ -48,5 +48,5 @@ int						ft_atoi(const char *str)
 	else if (a > 9223372036854775806 && sign == 1)
 		return (-1);
 	else
-		return ((int)(a * sign));
+		return (a * sign);
 }
